+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
-All notable changes to the "L13 Projects" extension will be documented in this file.
+All notable changes to the "Projects" extension will be documented in this file.
+
+## [0.14.0] - 2020-07-05
+
+### Added
+- Added welcome view to empty workspaces view.
+- Added save dialog to unknown workspace.
+- Added `l13Projects.sortWorkspacesBy` to group and sort workspaces.
+- Added `l13Projects.initialWorkspacesGroupState` to set the initial state of the workspace groups.
+- Added keyboard shortcuts for workspace slots `Ctrl/Cmd + L Ctrl/Cmd + 1 .. 9`.
+- Added keyboard shortcut for previous workspace `Ctrl/Cmd + L Ctrl/Cmd + 0`.
+
+### Changed
+- Changed display name from `L13 Diff` to `Diff Folders`.
+- Changed display name from `L13 Projects` to `Projects`.
 
 ## [0.13.2] - 2019-10-20